Day 25: fully solve it automatically
No need for manual inspection of a graph diagram. Find the shortest path between every pair of components and count the frequency of wires in those shortest paths. I figured that the most frequent wires would be the ones bridging the two dense networks.

The code is still far from satisfactory and could do with further simplification.